fix: #WB2-1604, fix feedback PR add hooks and modify function names
